requirements.txt
pandas>=1.5
matplotlib>=3.5
scipy>=1.9
python-dotenv>=0.21